requirements.txt
# API Clients
anthropic>=0.39.0
google-generativeai>=0.8.0
elevenlabs>=1.0.0
google-api-python-client>=2.148.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.17

# LINE Integration
line-bot-sdk>=3.0.0

# Video Processing
moviepy>=2.1.1
pillow>=11.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.2
requests>=2.32.0
pydantic>=2.0.0
